refactor: move flat testing code into a new script
